requirements.txt
configparser
logging
python-telegram-bot